+++ frontend/hashcat.launcher/src/jsx/data/algorithms.js
@@ -1,5 +1,5 @@
 import data from "./data"
 
 export const getAlgorithms = () => {
-	return data.data.algorithms || [];
+	return data.data.algorithms || {};
 };