+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project (hashcat.launcher) will be documented in thi
 ### Added:
  - Directory watcher (to auto load new added files) (#67)
  - Option to save cap to hcwpax output to hashes
+ - An alert if hashcat is not found
 
 ### Changed:
  - Moved default extra arguments to general (#66)